// src/types/enum.ts
export enum PRIORITY {
    NOTHING = 'NOTHING',
    LOW = 'LOW',
    MEDIUM = 'MEDIUM',
    HIGH = 'HIGH'
}

export enum REPEAT {
    NOTHING = 'NOTHING',
    FIVE_MINUTES = 'FIVE MINUTES',
    THIRTY_MINUTES = 'THIRTY MINUTES',
    ONE_HOUR = 'ONE HOUR',
    DAILY = 'DAILY',
    WEEKLY = 'WEEKLY',
    MONTHLY = 'MONTHLY',
    // CUSTOM = 'CUSTOM',
}

export enum EARLY_REMINDER {
    NOTHING = 'NOTHING',
    BEFORE_FIVE_MINUTES = 'BEFORE 5 MINUTES',
    BEFORE_FIFTEEN_MINUTES = 'BEFORE 15 MINUTES',
    BEFORE_THIRTY_MINUTES = 'BEFORE 30 MINUTES',
    BEFORE_ONE_HOUR = 'BEFORE 1 HOUR',
    BEFORE_ONE_DAY = 'BEFORE 1 DAY',
    BEFORE_ONE_WEEK = 'BEFORE 1 WEEK',
    BEFORE_ONE_MONTH = 'BEFORE 1 MONTH',
    // CUSTOM = 'CUSTOM',
}

export enum NOTIFICATION_TYPE {
    EARLY = 'EARLY',
    REPEAT = 'REPEAT',
}